address PR review, add quick ref (when no command)
