+++ templates/data/options.py
@@ -27,7 +27,7 @@ debug = False
 owners_uids = (200102491231092736,)
 
 # The extensions to load when running the bot.
-exts = ["ae7q", "base", "fun", "grid", "ham", "image", "lookup", "morse", "qrz", "study", "weather", "conv"]
+exts = ["ae7q", "base", "fun", "grid", "ham", "image", "lookup", "morse", "qrz", "study", "weather", "dbconv"]
 
 # Either "time", "random", or "fixed" (first item in statuses)
 status_mode = "fixed"
